Wrote new draft of the test plan.
Schedule files are included in /doc. The .numbers file is an OSX file for iWork
There's an xls file as well for you poor windows users :)

As far as I can see, the system testing plan needs be rewritten. I copied it
from a previous report, but I'm not too sure how to relate it our project.

I also didn't include a test case template and report template. I'm not sure
if these are required. If someone else wants to take the lead and just do it,
then I'll be ok with it. I'm more concerned with getting the text and everything else in order.